fixup: [AArch64][PAC] Lower code references to llvm.ptrauth globals.
Original message:

[AArch64] Lower null addr-disc to xzr before passing it to MOVaddrPAC.

We moved the discriminator computation to MOVaddrPAC/LOADgotPAC for
hardening.  But we passed the address discriminator directly to the
new nodes.

In the common case, a null value will end up getting optimized to an XZR
use directly in the pseudos, which is how the pseudo expansion
recognized the absence of an address discriminator.

However, in some cases, the null value survived in a vreg.  We should
really lower it directly to a target representation here.

We've used target constants elsewhere, but the addr disc is a register
in the cases where it's there: use XZR directly.
GlobalISel already uses XZR.

rdar://73407180