[AArch64][MachO] Lower @AUTH to ARM64_RELOC_AUTHENTICATED_POINTER.
This adds MachO support for emission of authenticated pointer
expressions.

In assembly, this is expressed using:
  .quad <symbol>@AUTH(<key>, <discriminator> [, addr])
For example:
  .quad _g3@AUTH(ib, 1234, addr)

Support for the @AUTH expression has already been generalized
into a separate commit.

The @AUTH expression lowers to a new kind of MachO relocation:
  ARM64_RELOC_AUTHENTICATED_POINTER
It points to the referenced symbol.

The other data, describing the signing scheme and original addend
(only 32 bits instead of 64), is encoded into the addend (in the
relocated location), as such:

  |63|62|61-51|50-49|  48  |47     -     32|31  -  0|
  | 1| 0|  0  | key | addr | discriminator | addend |