[clang] Implement the ptrauth_struct type attribute.
FIXME: the ptrauth-struct-attr.cpp/.m tests regressed
after 7eca38ce76d5d, which uses CreateGEP (that eagerly
emits a GEP instr to put a nuw flag on it, which in turn
causes the null-checked resign to be emitted as the raw address),
whereas it previously used CreateInBoundsGEP (which lazily
emits the GEP when it's needed, by accumulating the offset
into the Address, thus allowing the later dereference use
to mark the Address as known-non-null).

I have hacks to fix it, and we can obviously just update the test
to include the null-checks in resigns, but we should come up with
a long term solution.